+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -194,6 +194,14 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             toggleButtonsBorderSchemeColor:
                 controller.toggleButtonsBorderSchemeColor,
             toggleButtonsBorderWidth: controller.toggleButtonsBorderWidth,
+            // SegmentedButtons settings.
+            segmentedButtonRadius: controller.segmentedButtonBorderRadius,
+            segmentedButtonSchemeColor: controller.segmentedButtonSchemeColor,
+            segmentedButtonUnselectedSchemeColor:
+                controller.segmentedButtonUnselectedSchemeColor,
+            segmentedButtonBorderSchemeColor:
+                controller.segmentedButtonBorderSchemeColor,
+            segmentedButtonBorderWidth: controller.segmentedButtonBorderWidth,
             // Switch, CheckBox, Radio (toggles) settings.
             switchSchemeColor: controller.switchSchemeColor,
             switchThumbSchemeColor: controller.switchThumbSchemeColor,