+++ example/lib/example5_themes_playground/widgets/panels/panel_item.dart
@@ -94,8 +94,8 @@ const List<PanelItem> panelItems = <PanelItem>[
     icon: Icons.crop_16_9_outlined,
   ),
   PanelItem(
-    panelLabel: 'Toggle Buttons',
-    buttonLabel: 'Toggle\nButtons',
+    panelLabel: 'Selection Buttons',
+    buttonLabel: 'Selection\nButtons',
     icon: Icons.view_week_outlined,
   ),
   PanelItem(